Refactor and enhance ENABLE_MACRO_BLOCK_BLOOM_FILTER and ORGANIZATION support
