+++ libs/ob-sql-parser/src/test/java/com/oceanbase/tools/sqlparser/OBOracleSQLParserTest.java
@@ -66,40 +66,6 @@ import com.oceanbase.tools.sqlparser.statement.update.UpdateAssign;
 public class OBOracleSQLParserTest {
 
     @Test
-    public void parse_enableMacroBlockBloomFilterEqualsFalse_parseSucceed() {
-        SQLParser parser = new OBOracleSQLParser();
-        Statement actual = parser
-                .parse(new StringReader("CREATE TABLE \"EMPLOYEES\" (\n" +
-                        "  \"EMPLOYEE_ID\" NUMBER(10),\n" +
-                        "  \"FIRST_NAME\" VARCHAR2(50),\n" +
-                        "  \"LAST_NAME\" VARCHAR2(50),\n" +
-                        "  \"HIRE_DATE\" DATE,\n" +
-                        "  \"SALARY\" NUMBER(10,2),\n" +
-                        "  PRIMARY KEY (\"EMPLOYEE_ID\")\n" +
-                        ") COMPRESS FOR ARCHIVE REPLICA_NUM = 1 BLOCK_SIZE = 16384 USE_BLOOM_FILTER = FALSE " +
-                        "ENABLE_MACRO_BLOCK_BLOOM_FILTER = FALSE TABLET_SIZE = 134217728 PCTFREE = 0 "));
-        Assert.assertEquals(new Boolean("false"),
-                ((CreateTable) actual).getTableOptions().getEnableMacroBlockBloomFilter());
-    }
-
-    @Test
-    public void parse_enableMacroBlockBloomFilterEqualsTrue_parseSucceed() {
-        SQLParser parser = new OBOracleSQLParser();
-        Statement actual = parser
-                .parse(new StringReader("CREATE TABLE \"EMPLOYEES\" (\n" +
-                        "  \"EMPLOYEE_ID\" NUMBER(10),\n" +
-                        "  \"FIRST_NAME\" VARCHAR2(50),\n" +
-                        "  \"LAST_NAME\" VARCHAR2(50),\n" +
-                        "  \"HIRE_DATE\" DATE,\n" +
-                        "  \"SALARY\" NUMBER(10,2),\n" +
-                        "  PRIMARY KEY (\"EMPLOYEE_ID\")\n" +
-                        ") COMPRESS FOR ARCHIVE REPLICA_NUM = 1 BLOCK_SIZE = 16384 USE_BLOOM_FILTER = FALSE " +
-                        "ENABLE_MACRO_BLOCK_BLOOM_FILTER = TRUE TABLET_SIZE = 134217728 PCTFREE = 0 "));
-        Assert.assertEquals(new Boolean("true"),
-                ((CreateTable) actual).getTableOptions().getEnableMacroBlockBloomFilter());
-    }
-
-    @Test
     public void parse_selectStatement_parseSucceed() {
         SQLParser parser = new OBOracleSQLParser();
         Statement actual = parser.parse(new StringReader("select col.* abc from dual"));
