+++ pom.xml
@@ -94,7 +94,7 @@
         <bucket4j-core.version>4.10.0</bucket4j-core.version>
         <springfox-swagger-ui.version>2.10.0</springfox-swagger-ui.version>
         <db-browser.version>1.2.2</db-browser.version>
-        <ob-sql-parser.version>1.4.0</ob-sql-parser.version>
+        <ob-sql-parser.version>1.4.1</ob-sql-parser.version>
         <pf4j.version>3.10.0</pf4j.version>
         <bcpkix-jdk15on.version>1.64</bcpkix-jdk15on.version>
         <alibaba-dingtalk-service-sdk.version>2.0.0</alibaba-dingtalk-service-sdk.version>
